1. equipmentInfo 수정

+++ IdleGameServer/src/main/java/com/onlyonegames/eternalfantasia/domain/model/entity/MyEquipmentInfo.java
@@ -17,19 +17,19 @@ public class MyEquipmentInfo extends BaseTimeEntity {
     @GeneratedValue(strategy = GenerationType.TABLE, generator = "hibernate_sequence")
     Long id;
     Long useridUser;
-    public Long classEquipment;
+    public int classEquipment;
     public int nowUsedWeapon;
-    public Long swordEquipment;
-    public Long daggerEquipment;
-    public Long spearEquipment;
-    public Long bowEquipment;
-    public Long wandEquipment;
-    public Long earringEquipment;
-    public Long necklaceEquipment;
-    public Long ringEquipment;
+    public int swordEquipment;
+    public int daggerEquipment;
+    public int spearEquipment;
+    public int bowEquipment;
+    public int wandEquipment;
+    public int earringEquipment;
+    public int necklaceEquipment;
+    public int ringEquipment;
 
     public void SetClassEquipment(String element) {
-        this.classEquipment = Long.parseLong(element);
+        this.classEquipment = Integer.parseInt(element);
     }
 
     public void SetNowUsedWeapon(String element) {
@@ -37,34 +37,34 @@ public class MyEquipmentInfo extends BaseTimeEntity {
     }
 
     public void SetSwordEquipment(String element) {
-        this.swordEquipment = Long.parseLong(element);
+        this.swordEquipment = Integer.parseInt(element);
     }
 
     public void SetDaggerEquipment(String element) {
-        this.daggerEquipment = Long.parseLong(element);
+        this.daggerEquipment = Integer.parseInt(element);
     }
 
     public void SetSpearEquipment(String element) {
-        this.spearEquipment = Long.parseLong(element);
+        this.spearEquipment = Integer.parseInt(element);
     }
 
     public void SetBowEquipment(String element) {
-        this.bowEquipment = Long.parseLong(element);
+        this.bowEquipment = Integer.parseInt(element);
     }
 
     public void SetWandEquipment(String element) {
-        this.wandEquipment = Long.parseLong(element);
+        this.wandEquipment = Integer.parseInt(element);
     }
 
     public void SetEarringEquipment(String element) {
-        this.earringEquipment = Long.parseLong(element);
+        this.earringEquipment = Integer.parseInt(element);
     }
 
     public void SetNecklaceEquipment(String element) {
-        this.necklaceEquipment = Long.parseLong(element);
+        this.necklaceEquipment = Integer.parseInt(element);
     }
 
     public void SetRingEquipment(String element) {
-        this.ringEquipment = Long.parseLong(element);
+        this.ringEquipment = Integer.parseInt(element);
     }
 }
